ending album and photo

+++ src/main/java/ru/mrcrross/vphotoalbum/modules/photos/controllers/PhotoController.java
@@ -6,8 +6,10 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import ru.mrcrross.vphotoalbum.models.User;
 import ru.mrcrross.vphotoalbum.modules.photos.services.PhotoAlbumService;
 import ru.mrcrross.vphotoalbum.modules.photos.services.PhotoService;
+import ru.mrcrross.vphotoalbum.modules.user.services.UserService;
 import ru.mrcrross.vphotoalbum.wrappers.ControllerWrapper;
 
 @Controller
@@ -15,15 +17,16 @@ import ru.mrcrross.vphotoalbum.wrappers.ControllerWrapper;
 public class PhotoController extends ControllerWrapper {
     private final PhotoService photoService;
     private final PhotoAlbumService albumService;
+    private final UserService userService;
 
-    public PhotoController(PhotoService photoService, PhotoAlbumService albumService)
+    public PhotoController(PhotoService photoService, PhotoAlbumService albumService, UserService userService)
     {
         this.photoService = photoService;
         this.albumService = albumService;
+        this.userService = userService;
     }
     @GetMapping("")
-    public String index(HttpSession session)
-    {
+    public String index(HttpSession session) {
         if (this.loginControl(session)) {
             return "redirect:/";
         }
@@ -39,7 +42,11 @@ public class PhotoController extends ControllerWrapper {
         if (this.loginControl(session)) {
             return "redirect:/";
         }
+        if (photoService.checkViewer(id, (User)session.getAttribute("user"))) {
+            return "redirect:/";
+        }
         model.addAttribute("photo", photoService.getByID(id));
+        model.addAttribute("users", userService.getAll());
         return "views/photo/photo";
     }
 
@@ -52,7 +59,11 @@ public class PhotoController extends ControllerWrapper {
         if (this.loginControl(session)) {
             return "redirect:/";
         }
+        if (albumService.checkViewer(id, (User)session.getAttribute("user"))) {
+            return "redirect:/";
+        }
         model.addAttribute("album", albumService.getByID(id));
+        model.addAttribute("users", userService.getAll());
         return "views/photo/album";
     }
 
@@ -62,6 +73,9 @@ public class PhotoController extends ControllerWrapper {
             HttpSession session
     ) {
         if (this.loginControl(session)) {
+            return "redirect:/";
+        }
+        if (photoService.checkOwner(id, (User)session.getAttribute("user"))) {
             return "redirect:/";
         }
         photoService.delete(id);
@@ -76,7 +90,20 @@ public class PhotoController extends ControllerWrapper {
         if (this.loginControl(session)) {
             return "redirect:/";
         }
+        if (albumService.checkOwner(id, (User)session.getAttribute("user"))) {
+            return "redirect:/";
+        }
         albumService.delete(id);
         return "redirect:/photo";
+    }
+
+    @GetMapping(path = "viewer")
+    public String getViewerAlbums(
+            HttpSession session
+    ) {
+        if (this.loginControl(session)) {
+            return "redirect:/";
+        }
+        return "views/photo/viewer";
     }
 }
